Classify: Make internal DoClassify
In order to still have a helper function that is generic over the same
variable (classifier) as the ClassifyFn trait, use DoClassify as an
empty phantom structure which implements the do_classify function.
do_classify returns a Link rather than the DoClassify struct.

This helps flatten the library and conform to the expected layout
without pushing complexity too close to the UI portions of the code.