Create reworked Join
Join is now a function that takes a vector of links of the same type and
combines them together into a single link.  Future work may include
making a wrapper struct around vectors of links, so that join can be a
member function, rather than a standalone function. But this is an easy
way to collapse a bunch of Links back together.