Unzip: Serial only
Create unzip function, it uses a classify link to round robin a stream
into n number of parallel streams. Ideally, then the user can call
process on all of them. Had to rework classify a little, but I think
this new version is more flexible than the last. It comes with
mutability and the number of output streams is no longer a const item,
so that the number of output streams may be set at construction.